# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TransityConsole API - OTA aggregation & operator management
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: operators
    description: Operator registry management
  - name: terminals
    description: Terminal health monitoring
  - name: trips
    description: Aggregated trip search across operators
  - name: bookings
    description: Booking tracking across operators
  - name: analytics
    description: Analytics and reporting
  - name: gateway
    description: Gateway aggregation — proxy to TransityTerminal instances
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /operators:
    get:
      operationId: listOperators
      tags: [operators]
      summary: List all operators
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of operators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorListResponse"
    post:
      operationId: createOperator
      tags: [operators]
      summary: Register a new operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOperatorBody"
      responses:
        "201":
          description: Operator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"

  /operators/{id}:
    get:
      operationId: getOperator
      tags: [operators]
      summary: Get operator by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Operator details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
    patch:
      operationId: updateOperator
      tags: [operators]
      summary: Update operator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOperatorBody"
      responses:
        "200":
          description: Operator updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
    delete:
      operationId: deleteOperator
      tags: [operators]
      summary: Delete operator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /operators/{id}/ping:
    post:
      operationId: pingOperatorTerminal
      tags: [operators]
      summary: Ping operator terminal health
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ping result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalPingResult"

  /terminals/health:
    get:
      operationId: listTerminalHealth
      tags: [terminals]
      summary: Get health status of all registered terminals
      responses:
        "200":
          description: Terminal health overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalHealthOverview"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings across operators
      parameters:
        - name: operatorId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"

  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get overall analytics summary
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"

  /analytics/operators:
    get:
      operationId: getOperatorAnalytics
      tags: [analytics]
      summary: Get per-operator analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
      responses:
        "200":
          description: Per-operator analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperatorAnalytics"

  /analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [analytics]
      summary: Get revenue analytics over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
      responses:
        "200":
          description: Revenue analytics data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenuePeriod"

  /gateway/trips/search:
    get:
      operationId: gatewaySearchTrips
      tags: [gateway]
      summary: Aggregated trip search across all operators
      parameters:
        - name: originCity
          in: query
          required: true
          schema:
            type: string
        - name: destinationCity
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: passengers
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Aggregated search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewaySearchResult"

  /gateway/trips/{tripId}:
    get:
      operationId: gatewayGetTrip
      tags: [gateway]
      summary: Get trip details from a specific operator
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Trip not found

  /gateway/trips/{tripId}/seatmap:
    get:
      operationId: gatewayGetSeatmap
      tags: [gateway]
      summary: Get seatmap from a specific operator terminal
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: originSeq
          in: query
          required: true
          schema:
            type: integer
        - name: destinationSeq
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seatmap data
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Seatmap not found (trip may be virtual)

  /gateway/trips/{tripId}/reviews:
    get:
      operationId: gatewayGetReviews
      tags: [gateway]
      summary: Get reviews for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trip reviews
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Reviews not found

  /gateway/cities:
    get:
      operationId: gatewayGetCities
      tags: [gateway]
      summary: Aggregated cities from all operators
      responses:
        "200":
          description: Cities list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayCitiesResult"

  /gateway/operators/{operatorSlug}/info:
    get:
      operationId: gatewayGetOperatorInfo
      tags: [gateway]
      summary: Get operator brand info from terminal
      parameters:
        - name: operatorSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Operator info
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Operator not found

  /gateway/service-lines:
    get:
      operationId: gatewayGetServiceLines
      tags: [gateway]
      summary: Aggregated service lines from all operators
      responses:
        "200":
          description: Service lines
          content:
            application/json:
              schema:
                type: object

  /gateway/bookings:
    post:
      operationId: gatewayCreateBooking
      tags: [gateway]
      summary: Create booking via operator terminal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GatewayBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayBookingResult"
        "400":
          description: Validation error
        "503":
          description: Terminal unavailable

  /gateway/bookings/{bookingId}:
    get:
      operationId: gatewayGetBooking
      tags: [gateway]
      summary: Get booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Booking not found

  /gateway/payments/webhook:
    post:
      operationId: gatewayPaymentWebhook
      tags: [gateway]
      summary: Forward payment confirmation to operator terminal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GatewayWebhookPayload"
      responses:
        "200":
          description: Webhook forwarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayWebhookResult"
        "404":
          description: Booking not found for providerRef

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Operator:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        apiUrl:
          type: string
        serviceKey:
          type: string
        active:
          type: boolean
        logoUrl:
          type: string
          nullable: true
        commissionPct:
          type: number
        primaryColor:
          type: string
          nullable: true
        hasWebhookSecret:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - apiUrl
        - serviceKey
        - active
        - commissionPct
        - createdAt
        - updatedAt

    CreateOperatorBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        apiUrl:
          type: string
        serviceKey:
          type: string
        logoUrl:
          type: string
          nullable: true
        commissionPct:
          type: number
          default: 0
        primaryColor:
          type: string
          nullable: true
        webhookSecret:
          type: string
          nullable: true
      required:
        - name
        - slug
        - apiUrl
        - serviceKey

    UpdateOperatorBody:
      type: object
      properties:
        name:
          type: string
        apiUrl:
          type: string
        serviceKey:
          type: string
        active:
          type: boolean
        logoUrl:
          type: string
          nullable: true
        commissionPct:
          type: number
        primaryColor:
          type: string
          nullable: true
        webhookSecret:
          type: string
          nullable: true

    OperatorListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Operator"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean
      required:
        - data
        - total
        - page
        - limit
        - hasMore

    TerminalPingResult:
      type: object
      properties:
        operatorId:
          type: string
        status:
          type: string
          enum: [online, offline, degraded]
        latencyMs:
          type: number
          nullable: true
        checkedAt:
          type: string
          format: date-time
      required:
        - operatorId
        - status
        - checkedAt

    TerminalHealthEntry:
      type: object
      properties:
        operatorId:
          type: string
        operatorName:
          type: string
        operatorSlug:
          type: string
        status:
          type: string
          enum: [online, offline, degraded]
        latencyMs:
          type: number
          nullable: true
        lastCheckedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - operatorId
        - operatorName
        - operatorSlug
        - status

    TerminalHealthOverview:
      type: object
      properties:
        total:
          type: integer
        online:
          type: integer
        offline:
          type: integer
        degraded:
          type: integer
        terminals:
          type: array
          items:
            $ref: "#/components/schemas/TerminalHealthEntry"
      required:
        - total
        - online
        - offline
        - degraded
        - terminals

    Booking:
      type: object
      properties:
        id:
          type: string
        operatorId:
          type: string
        operatorName:
          type: string
        passengerName:
          type: string
        passengerPhone:
          type: string
        tripId:
          type: string
        origin:
          type: string
        destination:
          type: string
        departureDate:
          type: string
          format: date
        seatNumbers:
          type: array
          items:
            type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - operatorId
        - operatorName
        - passengerName
        - passengerPhone
        - tripId
        - origin
        - destination
        - departureDate
        - seatNumbers
        - totalAmount
        - status
        - createdAt

    BookingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean
      required:
        - data
        - total
        - page
        - limit
        - hasMore

    AnalyticsSummary:
      type: object
      properties:
        totalOperators:
          type: integer
        activeOperators:
          type: integer
        onlineTerminals:
          type: integer
        totalBookings:
          type: integer
        bookingsToday:
          type: integer
        totalRevenue:
          type: number
        revenueToday:
          type: number
        avgLatencyMs:
          type: number
          nullable: true
      required:
        - totalOperators
        - activeOperators
        - onlineTerminals
        - totalBookings
        - bookingsToday
        - totalRevenue
        - revenueToday

    OperatorAnalytics:
      type: object
      properties:
        operatorId:
          type: string
        operatorName:
          type: string
        operatorSlug:
          type: string
        bookingCount:
          type: integer
        revenue:
          type: number
        commissionEarned:
          type: number
        avgLatencyMs:
          type: number
          nullable: true
        uptimePct:
          type: number
      required:
        - operatorId
        - operatorName
        - operatorSlug
        - bookingCount
        - revenue
        - commissionEarned
        - uptimePct

    RevenuePeriod:
      type: object
      properties:
        date:
          type: string
          format: date
        revenue:
          type: number
        commission:
          type: number
        bookingCount:
          type: integer
      required:
        - date
        - revenue
        - commission
        - bookingCount

    GatewayTripStop:
      type: object
      properties:
        stopId:
          type: string
        cityName:
          type: string
        stopName:
          type: string
        sequence:
          type: integer
        departureTime:
          type: string
          nullable: true
        arrivalTime:
          type: string
          nullable: true
      required:
        - stopId
        - cityName
        - stopName
        - sequence

    GatewayTrip:
      type: object
      properties:
        tripId:
          type: string
        operatorId:
          type: string
        operatorName:
          type: string
        operatorSlug:
          type: string
        operatorLogo:
          type: string
          nullable: true
        operatorColor:
          type: string
          nullable: true
        serviceDate:
          type: string
          format: date
        origin:
          $ref: "#/components/schemas/GatewayTripStop"
        destination:
          $ref: "#/components/schemas/GatewayTripStop"
        farePerPerson:
          type: integer
        availableSeats:
          type: integer
        isVirtual:
          type: boolean
        vehicleClass:
          type: string
          nullable: true
      required:
        - tripId
        - operatorId
        - operatorName
        - operatorSlug
        - serviceDate
        - origin
        - destination
        - farePerPerson
        - availableSeats
        - isVirtual

    GatewaySearchResult:
      type: object
      properties:
        trips:
          type: array
          items:
            $ref: "#/components/schemas/GatewayTrip"
        errors:
          type: array
          items:
            type: object
            properties:
              operatorSlug:
                type: string
              error:
                type: string
        totalOperators:
          type: integer
        respondedOperators:
          type: integer
      required:
        - trips
        - errors
        - totalOperators
        - respondedOperators

    GatewayCitiesResult:
      type: object
      properties:
        cities:
          type: array
          items:
            type: string
        byOperator:
          type: array
          items:
            type: object
            properties:
              operatorSlug:
                type: string
              cities:
                type: array
                items:
                  type: string
      required:
        - cities
        - byOperator

    GatewayPassengerInput:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        idNumber:
          type: string
        seatNo:
          type: string
      required:
        - fullName
        - seatNo

    GatewayBookingRequest:
      type: object
      properties:
        tripId:
          type: string
        serviceDate:
          type: string
          format: date
        originStopId:
          type: string
        destinationStopId:
          type: string
        originSeq:
          type: integer
        destinationSeq:
          type: integer
        passengers:
          type: array
          items:
            $ref: "#/components/schemas/GatewayPassengerInput"
        paymentMethod:
          type: string
          enum: [qr, ewallet, bank]
      required:
        - tripId
        - serviceDate
        - originStopId
        - destinationStopId
        - originSeq
        - destinationSeq
        - passengers
        - paymentMethod

    GatewayBookingResult:
      type: object
      properties:
        bookingId:
          type: string
        externalBookingId:
          type: string
          nullable: true
        operatorId:
          type: string
        operatorName:
          type: string
        operatorSlug:
          type: string
        status:
          type: string
        totalAmount:
          type: string
        holdExpiresAt:
          type: string
          format: date-time
          nullable: true
        paymentIntent:
          type: object
          nullable: true
        qrData:
          type: array
          nullable: true
        passengers:
          type: array
        tripId:
          type: string
      required:
        - bookingId
        - operatorId
        - operatorName
        - operatorSlug
        - status
        - totalAmount
        - tripId

    GatewayWebhookPayload:
      type: object
      properties:
        providerRef:
          type: string
        status:
          type: string
          enum: [success, failed]
      required:
        - providerRef
        - status

    GatewayWebhookResult:
      type: object
      properties:
        success:
          type: boolean
        bookingId:
          type: string
        newStatus:
          type: string
      required:
        - success
        - bookingId
        - newStatus