# packages/api-zod/src/generated/types/gatewayBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TransityConsole API - OTA aggregation & operator management
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayBookingRequestPaymentMethod } from "./gatewayBookingRequestPaymentMethod";
import type { GatewayPassengerInput } from "./gatewayPassengerInput";

export interface GatewayBookingRequest {
  tripId: string;
  serviceDate: Date;
  originStopId: string;
  destinationStopId: string;
  originSeq: number;
  destinationSeq: number;
  passengers: GatewayPassengerInput[];
  paymentMethod: GatewayBookingRequestPaymentMethod;
}