# packages/api-zod/src/generated/types/gatewayCitiesResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TransityConsole API - OTA aggregation & operator management
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayCitiesResultByOperatorItem } from "./gatewayCitiesResultByOperatorItem";

export interface GatewayCitiesResult {
  cities: string[];
  byOperator: GatewayCitiesResultByOperatorItem[];
}