# packages/api-zod/src/generated/types/gatewayPassengerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TransityConsole API - OTA aggregation & operator management
 * OpenAPI spec version: 0.1.0
 */

export interface GatewayPassengerInput {
  fullName: string;
  phone?: string;
  idNumber?: string;
  seatNo: string;
}