# packages/api-zod/src/generated/types/gatewaySearchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TransityConsole API - OTA aggregation & operator management
 * OpenAPI spec version: 0.1.0
 */
import type { GatewaySearchResultErrorsItem } from "./gatewaySearchResultErrorsItem";
import type { GatewayTrip } from "./gatewayTrip";

export interface GatewaySearchResult {
  trips: GatewayTrip[];
  errors: GatewaySearchResultErrorsItem[];
  totalOperators: number;
  respondedOperators: number;
}