# packages/api-zod/src/generated/types/gatewayTrip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TransityConsole API - OTA aggregation & operator management
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayTripStop } from "./gatewayTripStop";

export interface GatewayTrip {
  tripId: string;
  operatorId: string;
  operatorName: string;
  operatorSlug: string;
  operatorLogo?: string | null;
  operatorColor?: string | null;
  serviceDate: Date;
  origin: GatewayTripStop;
  destination: GatewayTripStop;
  farePerPerson: number;
  availableSeats: number;
  isVirtual: boolean;
  vehicleClass?: string | null;
}